refactor: move pypi downloads out of deps-dev
