+++ services/apps/packages_worker/src/deps-dev/workflows/index.ts
@@ -3,6 +3,5 @@ export * from './ingestAdvisories'
 export * from './ingestDependentCounts'
 export * from './ingestDependencies'
 export * from './ingestPackages'
-export * from './ingestPypiDownloads'
 export * from './ingestRepos'
 export * from './ingestVersions'
